Created all of the unit tests for the ClaimRepository methods

+++ ClaimsChallenge/ClaimsLibrary/Claim.cs
@@ -9,7 +9,7 @@ namespace ClaimsLibrary
         private static int TotalClaims = 0;
 
         //Constructor
-        public Claim(ClaimType claimType, string description, decimal claimAmount, DateTime dateOfIncident, DateTime dateOfClaim)
+        public Claim(ClaimType claimType, string description, double claimAmount, DateTime dateOfIncident, DateTime dateOfClaim)
         {
             ClaimId = ++TotalClaims; //++ has to be before because it will increment before taking action, will set to 1 first and then assign, instead of the other way around
             ClaimType = claimType;
@@ -25,7 +25,7 @@ namespace ClaimsLibrary
 
         public string Description { get; set; }
 
-        public decimal ClaimAmount { get; set; }
+        public double ClaimAmount { get; set; }
 
         public DateTime DateOfIncident { get; set; }
